Add basic sequence derivative functions
The problem with typeclass instances not being imported can be
reproduced by simply not typeclass terms in the top scope while
requiring them in a lower scope.

A second issue is with expressions like `init []`, where the parameter
of the list cannot be inferred since it is empty. Here I should either
allow the unused parameter to propagate (which is probably not OK) or
require an annotations. So it would be `init ([] :: [Int])`. But the
annotation approach is currently not working. This means something is
screwy in the typechecker.